requirements.txt
fastapi
uvicorn
requests
pandas
numpy
matplotlib
streamlit
openai
torch
azure-storage-file-datalake
azure-identity